pyproject.toml
[tool.poetry]
name = "todam-ticket-system"
version = "v0.0.1"
description = ""
authors = ["1chooo <hugo970217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.28.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
